Midterm Problem 1
Define and Implement a C++ class, i.e. TwoDVector, to represent a two dimensional vector, include the definition but not implementation of the following:

Default Constructor
Copy Constructor
Parametrized constructor
getX(), getY() getters
setX(), setY() setters
getTheta() to get the angle of the vectors with respect to x positive axis
Overload the << operator to pretty print the vector components X and Y
Overload the ! operator to reverse the direction of the vector
Overload the * operator to perform a dot product of the two vectors
Overload the + operator to add the components of the two vectors
Overload the - operator to subtract the components of the two vectors
Overload the / operator to divide a vector by a scalar and reduce the magnitude of the vectors by the scalar.
Overload the ++ and -- prefix operators to increase the magnitude of X and Y components by 1
Overload the () operator to return the magnitude of the vector