Problem 3
Recursion: Assume a child can take 1, 2, or 3 steps at a time. Given a stair of n steps, create a C++ recursive function to compute the number of possible ways the child could complete the stair. As an example a stair of 3 steps the child can use the following number of way: he can take only steps of 1: {1, 1, 1}. These are also possibilities: {1, 2}, {2, 1}, and {3}. So the number of ways to complete 3 steps is 4.